+++ packages/cli/src/cms/commands/push.ts
@@ -1,8 +1,9 @@
 import * as fs from 'fs';
 import * as path from 'path';
 import { slash } from '@redocly/openapi-core';
+import { pluralize } from '@redocly/openapi-core/lib/utils';
 import { green, yellow } from 'colorette';
-import { exitWithError, pluralize, printExecutionTime } from '../../utils/miscellaneous';
+import { exitWithError, printExecutionTime } from '../../utils/miscellaneous';
 import { handlePushStatus } from './push-status';
 import { ReuniteApiClient, getDomain, getApiKeys } from '../api';
 import { handleReuniteError } from './utils';