+++ packages/core/src/config/recommended-strict.ts
@@ -110,7 +110,17 @@ const recommendedStrict: PluginStyleguideConfig<'built-in'> = {
   },
   arazzoRules: {
     spec: 'error',
-    'parameters-no-body-inside-in': 'off',
+    'parameters-not-in-body': 'error',
+    'sourceDescription-type': 'error',
+    'version-enum': 'error',
+    'workflowId-unique': 'error',
+    'stepId-unique': 'error',
+    'sourceDescription-name-unique': 'error',
+    'workflow-dependsOn': 'error',
+    'parameters-unique': 'error',
+    'step-onSuccess-unique': 'error',
+    'step-onFailure-unique': 'error',
+    'requestBody-replacements-unique': 'error',
   },
 };