+++ packages/core/src/rules/arazzo/index.ts
@@ -1,6 +1,16 @@
 import { Spec } from '../common/spec';
 import { Assertions } from '../common/assertions';
-import { ParametersNoBodyInsideIn } from '../spot/parameters-no-body-inside-in';
+import { ParametersNotInBody } from '../spot/parameters-not-in-body';
+import { SourceDescriptionType } from '../arazzo/source-description-type';
+import { VersionEnum } from '../spot/version-enum';
+import { WorkflowIdUnique } from './workflowId-unique';
+import { StepIdUnique } from './stepId-unique';
+import { SourceDescriptionsNameUnique } from './sourceDescriptions-name-unique';
+import { WorkflowDependsOn } from './workflow-dependsOn';
+import { ParametersUnique } from './parameters-unique';
+import { StepOnSuccessUnique } from './step-onSuccess-unique';
+import { StepOnFailureUnique } from './step-onFailure-unique';
+import { RequestBodyReplacementsUnique } from './requestBody-replacements-unique';
 
 import type { ArazzoRule } from '../../visitors';
 import type { ArazzoRuleSet } from '../../oas-types';
@@ -8,7 +18,17 @@ import type { ArazzoRuleSet } from '../../oas-types';
 export const rules: ArazzoRuleSet<'built-in'> = {
   spec: Spec as ArazzoRule,
   assertions: Assertions as ArazzoRule,
-  'parameters-no-body-inside-in': ParametersNoBodyInsideIn as ArazzoRule,
+  'parameters-not-in-body': ParametersNotInBody as ArazzoRule,
+  'sourceDescription-type': SourceDescriptionType as ArazzoRule,
+  'version-enum': VersionEnum as ArazzoRule,
+  'workflowId-unique': WorkflowIdUnique as ArazzoRule,
+  'stepId-unique': StepIdUnique as ArazzoRule,
+  'sourceDescription-name-unique': SourceDescriptionsNameUnique as ArazzoRule,
+  'workflow-dependsOn': WorkflowDependsOn as ArazzoRule,
+  'parameters-unique': ParametersUnique as ArazzoRule,
+  'step-onSuccess-unique': StepOnSuccessUnique as ArazzoRule,
+  'step-onFailure-unique': StepOnFailureUnique as ArazzoRule,
+  'requestBody-replacements-unique': RequestBodyReplacementsUnique as ArazzoRule,
 };
 
 export const preprocessors = {};