+++ packages/core/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @redocly/openapi-core
 
+## 1.22.1
+
+### Patch Changes
+
+- Fixed an issue where resolving config in the browser always threw an error.
+
+## 1.22.0
+
+### Minor Changes
+
+- Updated the Arazzo validation types for workflows input, parameter objects, and criteria to match the specification.
+- Added Arazzo rulesets so that users can customize their linting rules for this format.
+
+### Patch Changes
+
+- Updated @redocly/config to v0.10.1.
+
 ## 1.21.1
 
 ### Patch Changes